Extract Full Model Execution Tests into Nightly Tests (#379)
### Ticket
#378 

### Problem description
Full model execution tests are only attempted in push/pr triggered
workflows, but not in nightlies. This complicates queries and provides
no single source of truth on actual full model execution status from 
tt-torch main.

### What's changed
Duplicate the workflow in run-tests.yml into a self contained
job that can be run in nightly tests.  Execution tests are refactored
to be run in parallel.

### Checklist
- [x] New/Existing tests provide coverage for changes