Add model report file name check to op by op parser (#419)
### Ticket
#418 

### Problem description
Op by op parser may fail to parse a JSON on CI rerun
because it is malformed. This occurs when reingesting
data from JSON generating jobs in the old run during 
a CI rerun that are not model test data.

### What's changed
Add tighter check to ensure we only parse model op
report JSONs.

### Checklist
- [x] New/Existing tests provide coverage for changes