Torchvision object detection model fails on torch compile (#432)
### Ticket
#261 

### Problem description
1. torchvision object detection model was failing during torch compiler
stage. When a torch subgraph is only made up of an output node we remove
it and return an empty subgraph, causing torch.fx to Fail during shape
propagation.
2. torchvision ssd model output is a List[Dict[Tensor]], we only
supported a structure of nested Lists/Tuples of Tensors, causing the
validation stage to Fail while trying to flatten the output

### What's changed
- Remove the last node == output node pruning on reduce_graph()
- Added a check for dict types, sort dict by key, make sure golden and
output have matching keys and flatten the tensors in the dict values
- Removed xfail from other models failing during compilation from the same issue

### Checklist
- [x] New/Existing tests provide coverage for changes