Fix Deepseek Qwen Fatal Bus Error [#291]
Deepseek Qwen op no. 9552 fails with out-of-memory error. This is
expected to be caught by tr-except block. However, it is not. Instead,
the whole program crashes.

The error is related to multiprocessing in Python. There is no full
isolation between parent and child processes. So, the child process
running out of memory kills the parent process. I deduced this by
running several tests:

- a single threaded `run_op` function - successfully passes
- a multi processed run_op function who only calls
  `tt_mlir.run_end_to_end`, does not do anything related to stderr
piping etc - fails

Further research showed that memory is not used in full isolation in
multiprocesses between parent and child.

Solution: replacing a multi-processed `run_op` with one that runs a
subprocess, since subprocesses provide complete isolation.