Use Torch ExportedModule to import initial MLIR module (#416)
Use a `torch.export.ExportedProgram` to generate the initial MLIR
module.

This requires us to create an `ExportedProgram` from the initial
`GraphModule`.

Benefits:
- We can use the torch-mlir's official entrypoint
    - This handles in-place ops for us
- We can run decompositions and **keep** location data
- This location data will stick around throughout the compile process

Issues:
- `aten.clamp` is decomposed by torch-mlir to `maximum(minimum(input,
max), min)`. `ttnn.maximum` requires that the operand which needs to be
broadcasted is on the RHS. Currently, in tt-mlir the
`PartiallyBroadcastable` op trait only enforces that the broadcasted
operand is on the LHS
    - tt-torch issue: https://github.com/tenstorrent/tt-torch/issues/431
    - tt-mlir issue: https://github.com/tenstorrent/tt-mlir/issues/2458
- Graph parameters are inlined as constants in the graph. To have the
`FxImporter` treat them as graph inputs we need to edit the
`ExportedModule`s `ExportedGraphSignature` and force all "parameter"
types to "user inputs"
- This is a hack as the `ExportedGraphSignature` is meant to be a
private member of `ExportedProgram`
- Ideally we can configure the `FxImporter` to _not_ inline the
parameters if we pass a flag of some sort. Perhaps a future contribution
to torch-mlir.

Other Info:
- We need to upgrade to PyTorch 2.6.0 as it contains crucial changes
which allow us to use custom decompositions (necessary to support
interpolation)
- AdaptiveAvgPool2d is lowered AvgPool2d and eventually to
`stablehlo.reduce_window **even in the case where the op is equivalent
to a global average**. Since we do not have support for lowering a
sum_pool in `StablehloToTTIRPatterns.cpp` (sum because the division is
afterward), I've temporarily added a custom decomposition of
`aten.avg_pool2d` which will convert to a mean over the spatial
dimensions when the `avg_pool2d` is equivalent to it.
- `aten.split` is no longer lowered to a series of `narrow` ops. Instead
it is now lowered to a series of `as_strided` ops.
- `narrow` is lowered to `slice`, which can be lowered to
`stablehlo.slice`. `as_strided` cannot be lowered from Torch Backend IR
to Stablehlo. I've temporarily added back the old decomposition from
PyTorch 2.5.0 which uses narrow as a custom decomposition.
- I've made a PR which adds a lowering of `AtenAsStridedOp` to
`stablehlo::SliceOp` in our fork of torch-mlir:
https://github.com/tenstorrent/llvm-torch-mlir/pull/4
- The tracer which generates the `GraphModule` which is passed to
`backend` does not account for control flow, I believe in PyTorch 2.5.0
a graph break would be triggered during `.generate` methods in
`transformers` LLMs. It does not anymore and so `.generate` will run
until the max length is reached.
- **this means that the entire generation becomes one program**
- Once the first EOS token is generated, the rest of the length is
filled with padding. We cannot compare the golden output to the result
from the `GraphModule` as the output shapes are different.
- Since the output of `.generate` graphs are integers PCC/atol
verification is not quite useful but does return `True` when the outputs
are _identical_
             - The tokenizer can decode the outputs and strip padding.
- I've added a flag to `ModelTester` that informs the `ModelTester` it
is testing a `.generate` call. It will decode the output tokens and we
compare the resulting strings.
- PyTorch has an experimental `torch.cond` which they seem to intend to
use to trace data-dependent control-flow. There's a note in the
`transformers` source that says they intend to use it when it is no
longer experimental
- When the graph is compiled, the user inputs are placed **at the end**
of the arguments passed to the program rather than the front. That is
graph constants first, then inputs.
- I needed to implement an `FxImporter` hook for importing literals to
the graph. By default it will make all non-scalars
`DenseElementsResourceAttr`s, however, this causes the process to hang
upon cleanup whether the test fails or not. So the hook just uses
`DenseElementsAttr` for all literals.
- Someone has mentioned this problem in an IREE issue as well:
https://github.com/iree-org/iree/issues/20102
- They've traced it down to this PR in llvm that adds a GIL acquire when
destroying the `DenseElementsResourceAttr`:
https://github.com/llvm/llvm-project/pull/124832