Isolate Full Model Execute Pytest Processes and Quarantine Bert (#562)
### Ticket
None

### Problem description
Full model pytests are run in a single invocation of pytest and
therefore generate a single test report and run in a single process.
Process-killing failures are rare, but will cause the report to not be
generated.

### What's changed
- Isolate full execute pytest processes so each generates individual
report
- Restore op-by-op and compile jobs temporarily removed earlier today

### Checklist
- [x] New/Existing tests provide coverage for changes
- [x] CollectData job picks up all pytest results even though they are
written to unique files. Post-processed report JSONs look well formed on
inspection and are queryable from Superset