Replace PCC/ATOL invalid numeric values with -1 (#573)
### Ticket
None

### Problem description
YoloV5 results are not being ingested by the database because the tags
it generates are not parsable. This is because it generates a numeric
infinite ATOL.

Both the JSON and postgres spec "officially" do not support NaN or
infinity values and those should be stripped from the reports.

### What's changed
Replace infinity and NaN values with -1 in PCC and ATOL. 

### Checklist
- [x] New/Existing tests provide coverage for changes