Improve Benchmark CI (Auto load balancer, out of band matrix autogeneration, test runner isolation) (#604)
### Ticket
None

### Problem description
Benchmarks are unreliable due to:
- The possibility of timeout, and the impossibility of setting a timeout
duration because we cannot know the a priori test runtime. Using
previous runtime will produce false negatives because runtime may
rightly drift week over week, and we'd have to requeue jobs into a new
workflow to avoid a global timeout to disambiguate timing out tests vs.
elongated tests.
- The possibility of device hang causing all tests beyond hanging point
in a group to fail even if we had a timeout-terminator without a device
reset mechanism
- The possibility of multiple timeout/hang causing a runner bottleneck
for remaining tests

### What's changed
- Isolate benchmark tests in their own runner job, so there is a clean
state before running each test. This will lead to problems with job
count, as we have a hard limit of 250, but haven't hit it yet
- Strip already EXECUTing tests from benchmarks as they are fully
redundant and are unpromotable anyways
- Add extra testname sanitation level to avoid artifact upload failure
as seen in previous runs
- Always run report generation even if previous steps fail for an
unforseen reason

### Checklist
- [x] Ran dummy benchmarks to check functionality