Add add_onnx_model_proto to stablehlo backend #543
Stablehlo Executor would use the binary from last op in op-by-op mode to
generate an output. However, this fails in cases where the last op
didn't compile or run. So, added support for running an onnx model proto
from the stablehlo executor. This entailed restructuring onnx code so
that there are 3 essential methods:

- prepare_inference_session: Created an onnx inference session and sets
  number of threads to use to avoid info messages clobbering the output.
- run_model_proto: runs an onnx model proto
- onnx_output_to_torch: standardized all backends to return a torch
  output, and this method helps convert onnx output (a numpy object) to
a torch object