Add add_onnx_model_proto to stablehlo backend #543 (#623)
### Ticket
#543 

### Problem description
When we are running an onnx model with stablehlo backend in an op-by-op
flow, we rely on the binary flatbuffer generated by the last op to
generate an output. However, this is unreliable in cases where last op
doesn't compile or run.

### What's changed
Added support for running an onnx model proto from the stablehlo
executor. This entailed restructuring onnx code so that there are 3
essential methods:
- `prepare_inference_session`: Created an onnx inference session and
sets number of threads to use to avoid info messages clobbering the
output.
- `run_model_proto`: runs an onnx model proto
- `onnx_output_to_torch`: standardized all backends to return a torch
output, and this method helps convert onnx output (a numpy object) to a
torch object

Added `add_onnx_model_proto` to `StablehloExecutor` which sets the
model_proto object.

Assumed that onnx output should be torch by default, thereby making
`_extract_outputs` not needed for casting numpy to torch.