Restore Model Caches at HF_HOME and TORCH_HOME (#630)
### Ticket
None

### Problem description
Fix bug causing HF_HOME and TORCH_HOME to resolve to paths inside the
container due to incorrect reference of environment variable. This
caused all model caches to be misdirected to an empty in-container path
at the start of job execution since April 9, explaining the increased
occurrence of CI failures due to network failures and rate limits (which
would have been avoided if a cache hit was possible).

### What's changed
Move DOCKER_CACHE_ROOT env variable to workflow-global scope. It is
still accessible as an env variable from the any shell, run in a step or
job within the workflow.

### Checklist
- [x] New/Existing tests provide coverage for changes