Disable golden calculation/verification when evaluating in op_by_op mode (#639)
### Ticket
Follow up on #329 and #637

### Problem description
When working on models from tickets like #329 and #637 on op_by_op mode
the compilation/debugging time is extended because we run the full torch
model to get the golden first. The golden is not needed for the op_by_op
flow so we want to disable it. This should also speed up the op_by_op CI
workflows.

### What's changed
Separated test_model_eval() into two modes:
- if the compile config depth is set to op_by_op we just compile and run
   the model using the op_by_op executors, and we skip the golden/output, 
- if not, we use the already existing logic to eval compilation and verify 
   golden

### Checklist
- [x] New/Existing tests provide coverage for changes