Fix data parallel demo by creating a new instance of the `options`  (#643)
### Problem description
In the initial multi-device
[PR](https://github.com/tenstorrent/tt-torch/pull/572) that added the
data parallel demo, there was a bug that caused the models to only
compile for the last device in the acquired `devices` list. This was due
to the options map being passed by reference to each torch.compile call.
This is fixed now by creating a fresh map for each iteration.

### Checklist
- [x] New/Existing tests provide coverage for changes