#580 Get system descriptor for opened device using updated API from mlir commit 4d13d92 (#628)
### Ticket
https://github.com/tenstorrent/tt-torch/issues/580

### Problem description
We couldn't create a system descriptor from tt-mlir while having the
device open before.
But now that the function `getCurrentSystemDesc` allows opened device to
be passed in, we can use it.
This will be useful for multidevice to make sure we pass system
descriptor of the device being used for that workload.

### What's changed
- Create a new system desc for every compilation.
- Pass in device being used to get system desc for that device.

### Checklist
- [ ] New/Existing tests provide coverage for changes
- [x] Make sure we generate system_desc when it's read and not later. To
prevent segfault from tt-mlir when wrong version of system_desc passed
in