Add llama 7b pipeline parallel test (#679)
### Ticket
closes #591 

### Problem description
We want a POC of a large model that doesn't fit on single device to run
in runtime pipeline parallel fashion on multiple chips.

### What's changed
Added llama 7b that is split in half and runs pipeline parallel on 2
chips.

Since this doesn't fit into our `ModelTester` flow and also since
LLama3b is getting bad pcc, adding this test currently with a separate
flow. One future task is to integrate this into our `ModelTester` flow
so we can properly gather results/reports like we do for other models
FYI @jameszianxuTT .

### Checklist
- [X] New/Existing tests provide coverage for changes