Enable typecast to graph constants
Workaround for int64 constant feeding issue on our device.
Our device misinterprets int64 tensors when they are provided as constants, resulting in like [1, 0, 2, 0, 3, 0, ...] instead of [1, 2, 3, ...].
To prevent this, this function safely casts int64 tensors to int32 if all values fit within the int32 range, ensuring correct data is passed to the device.