aknezevic/disable pruning (#791)
### Problem description
I recently added the option to delete unused constants for graphs. These
would come about if we had consteval enabled, as both the constevaled
and the original params were in the graph. This works for most tests,
but fails in compile only
[tests](https://github.com/tenstorrent/tt-torch/actions/runs/15081491402/job/42406680894),
since these actually need the original params (to execute the original
GM).

### What's changed
Only prune inputs for full execute tests. 

### Checklist
- [x] New/Existing tests provide coverage for changes