Fix Flux tests after Torch + Transformers Uplift #868
`pipeline` has problems with `torch.compile`. It is suggested to compile
sub-component of pipeline individually instead of the entire pipeline.
The new Flux tests compile `pipeline.transformer`

Reference tests for `schnell` and `dev` included in the original issue.

`dev` variant runs out-of-memory when running op-by-op, but can compile
end-to-end.

`schnell` variant is using high memory runner for op-by-op test.