Fix flux-dev out-of-memory #912 (#938)
### Ticket
#912 

### Problem description
`flux-dev` can compile end to end, but it gets killed when running
op-by-op, which hints a memory problem. The error did not happen at a
specific op.

### What's changed
The original test hadn't enabled consteval, and with consteval, the
number of ops in graph has greatly reduced. `flux-dev` does not get
killed anymore.