Fix clamp op decomposition
In  clamp op, `min` and `max` values are not casted to input data type correctly. Therefore, `max` is defaulted to `float32` whereas the input is `min`. This caused phi 3.5 vision to fail with the following error: `RuntimeError: Expected tensor for argument #1 'indices' to have one of the following scalar types: Long, Int; but got torch.FloatTensor instead (while checking arguments for embedding)`