Add path to disable backend consteval pass from compiler config (#951)
### Ticket
None

### Problem description
The default for the consteval option in compileTTIRToTTNN changed a
month ago to default to true. This caused our consteval environment
variable to be unable to suppress the backend consteval pass. We also
didn't have a programmatic path to set the TTIRToTTNN consteval pass
option without an env variable.

Required as a stopgap for #614 

### What's changed
Plumb the compiler_config.enable_consteval option through the backend to
allow it to be disabled properly, remove the now invalid use of the env
variable.

### Checklist
- [x] New/Existing tests provide coverage for changes