Add Falcon3 7B and 10B Pipeline Parallel Tests (#986)
### Ticket
https://github.com/tenstorrent/tt-torch/issues/922 

### Problem description
Add N300 pipeline parallel tests for falcon 3 models that don't fit on
one device.

### What's changed
Added pipeline parallel tests for falcon3 7B and 10B. 

**Issue**: The 10B model introduced a new error where
`infer_auto_device_map()` assigns modules to device `"disk"` when it
estimates devices 0 and 1 will be full. This causes
`split_onto_devices()` to fail since it expects only integer device
assignments. Additionally, manual changes to `device_map` proved that
this model can fit on 2 devices and so the prediction that device 0 and
1 will be full is not always correct. Each device has 12GiB of memory, but setting the max_memory to that amount causes an OOM error with llama's pipeline parallel test.

**Solution**: Change the max_memory to be 11GiB per device.


[EDIT]: A self contained path to creating a junitxml report for tests
that bypass the ModelTester was also added, which extracts the common
logic from ModelTester report generation and exposes it as a new static
function ModelTester.GenerateCustomTestReport

Summary:
- **tt-torch/backend/passes.py**
  - move modules assigned to `"disk"` to the last device
  - warn user that increasing max_memory might be necessary
- **tests/falcon/test_falcon3_7b_10b_pipeline_parallel.py**
- replicates format of
**tests/llama/test_llama_7b_pipeline_parallel.py**
  - Falcon3 10B `max_memory` increased to 10GiB per device
- **.github/workflows/run-full-model-execution-tests.yml**
- added tests to on-pr high-mem quarantine since they fail on lower
memory n300s (same as the llama test)
 

### Checklist
- [x] New/Existing tests provide coverage for changes

---------

Co-authored-by: James Zian Xu <jameszianxu@tenstorrent.com>