Move implicit executor device acquisition to start of executor __call__ (#1011)
### Ticket
None

### Problem description
Automatic device acquisition via _get_device() happens after cache
initialization, so if the device is not specified, the first executor
__call__ acquires a device during get_inputs(). This causes the initial
constant / buffer cache to be attached to a None device.

At the first cache insertion time, the device _is_ initialized by
automatic device acquisition, so the cache has two incoherent
perspectives on the device to attach the cache to - the list of tensors
to cache is attached to a None device while the actual allocated runtime
tensors are attached to a real device. The caching operation "misses"
due to device mismatch.

Then, at destruction time, a runtime error is encountered due to
attempted deallocation of uncached (None) runtime tensors.

### What's changed
- Move automatic device acquisition to the start of the Executor
__call__ so the device exists before cache initialization time.
- (minor) Add output filtering to llama demo so it only prints output
tokens to stdout.
### Checklist
- [x] New/Existing tests provide coverage for changes