Einnes tt/new torch instructions (#1028)
### Ticket
#1027 

### Problem description
We need clear, accurate instructions for how to use TT-Torch with a
wheel, with docker, and building from source

### What's changed
The old instructions moved to their own build page and focus on build
only. Wheel and Docker instructions are on their own pages with complete
steps for each. Docker uses the new container that was created for
TT-Torch.

### Checklist
- [ ] New/Existing tests provide coverage for changes