Microsoft/phi-4 & phi-3 variants bring up (#1023)
### Ticket
https://github.com/tenstorrent/tt-torch/issues/330

### Problem description
Tests for the Microsoft phi models, which are priority models, needed to
be added.

### What's changed
Through test_phi_4.py file we can see that phi-4 runs fully in op-by-op
and compiles fully down to TTNN_IR as well but does not execute
currently due to OOM issue. For that I've added a skip_eval_test.
The same applies to Phi-3-mini-4k-instruct and Phi-3-mini-128k-instruct.

CI Changes:
- Added phi-4 to op-by-op nightly and full-model-nightly under skip_red
category.
- Added the two phi-3 variants to full-model-nightly under skip_red
category.
- Added phi-3-mini-4k only to op-by-op-nightly as 4k-instruct and
128k-instruct only differ in token length and I didn't think it was
necessary to use resources to run the larger token length variant as it
takes more time.