Bring up pipeline parallel static cache model  (#1042)
### Ticket
#1010 

### Problem description
Add generative loop with static caching to llama_7b model which runs
using pipeline parallel. Combines **/demos/llama3.2_generate_demo.py**
with **/tests/models/llama/test_llama_7b_pipeline_parallel.py**. This
combination introduced mismatches in input sizes between
`example_inputs` and `gm.graph_inputs` before lowering as well as
incorrect labeling of input types (`USER` vs `INTER_DEVICE`) after
lowering.

### What's changed
The test **test_llama_7b_generative_pipeline_parallel.py** was created
and added to **run-full-model-execution-tests.yml** with the other
pipeline parallel tests.

The remaining changes are in **tt-torch/dynamo/passes.py**:

The `prev_device_idx` is set to the lowest device the previous node was
needed on instead of the highest.
- a placeholder (input) was being assigned to the second device when it
was need on the first. This created an unnecessary `INTER_DEVICE` input
and a mismatch with the `example_inputs` ordering. This was happening
because the input was assigned to the `prev_device_idx` which was the
second device since the previous input was needed on both devices.

The `user_input_index` (which sets the producer_index) only increments
once per node.
- when it incremented too many times the `producer_index` became
incorrect and so certain `gm.graph_inputs` were pointing to the wrong
input in `example_inputs` causing shape mismatches

The values in `node_to_new_nodes` aren't rewritten if the `device_idx`
is different
- when they were accidentally re-written there were nodes unnecessarily
added as `INTER_DEVICE` inputs which caused an issue since
`tt_mlir.run()` got more inputs than expected.

Added an assert to check that the input shape matches what's expected to
help catch any similar issues in the future.

### Checklist
- [x] New/Existing tests provide coverage for changes

### To-Do
When testing pipeline parallel on **llama3.2_generate_demo.py** the
device map was assigning the last module to the first device which
should cause all modules to move to the first device (bad). However,
`sort_device_map()` was not moving the modules over because it doesn't
handle `input_device=None` correctly which is something that needs to be
fixed. The test causing the issue and a more descriptive explanation is
on [this
branch](https://github.com/tenstorrent/tt-torch/tree/gengelage/issue-1010-llama3b-bug).
- [ ] Fix sort_device_map to handle input_device = None
- [ ] Consider adjusting sort_device_map to ensure that some modules
remain on later devices.

[EDIT]: issues added #1043 #1044