Enable PCC check for blackhole models (#1041)
### Ticket
[1003](https://github.com/tenstorrent/tt-torch/issues/1003)

### Problem description
For certain models using bfloat16 (not models that are entirely
bfloat16, but those with specific bfloat16 usage), there is an issue
with untilize on the device.

### What's changed
The workaround in [PR
#4024](https://github.com/tenstorrent/tt-mlir/pull/4024) applies running
untilize on the host. This commit resolves the issue for the following
models:
 - VGG series models
 - hardnet
 - dla34.in1k
 - yolov4

Also, there are two models with a slight difference compared to the
target PCC. We can enable a tolerance for these, and the target can be
updated for the BH architecture.
 - albert_masked_lm (pcc  = 0.9865)
 - resnet50 (pcc = 0.9762)

### Checklist
- [ ] New/Existing tests provide coverage for changes