Defer PCC/ATOL assertions for model tests until after tags reported (#1087)
### Ticket
Tests that assert on and then fail ATOL or PCC checks will not show up
in database reports. This causes them to be absent from a nightly
report, rather than showing up as explicitly failed, which is bad for
observability.

### Problem description
ATOL and PCC data (along with much metadata in the rest of `tags`) is
not flushed to the XML file in tester.finalize() if the assert_atol/pcc
assert triggers, because the AssertionError short circuits execution and
tester.finalize() cannot be afterwards, which will mess up test cleanup
(e.g. device closure) and not report.

### What's changed
Defer ATOL and PCC assertions for tests using the ModelTester to after
tester.finalize() is called, allowing tags to be gracefully flushed to
XML before the assertion takes over control flow.

This also allows us to xfail on PCC and ATOL, so tests can notify us
when they start passing due to uplift fixes.

### Checklist
- [x] Checked that a dummy testcase with failing ATOL/PCC assertions
does produce a report that is ingested and queryable. Models that failed
ATOL/PCC now show up, rather than being absent from the database
entirely.