Flip ModelTester default backend=tt-experimental and related test updates, remove experimental nightly CI (#1103)
### Ticket
None

### Problem description
- We want to switch to using tt-experimental (tt-xla) backend by default
across most tests sooner rather than later. Bunch of testing, pass rate
is high. Will continue to push on remining failures in Nightly if any.
- We also don't have CI capacity to maintain 2 nightlies (normal and
experimental) especially on BH side, so if pass rate is high prefer to
get this in to ease CI burden.

Note: The tenstorrent branch of torch-xla that we'd like to consume
isn't available on PyPi yet, but did some preliminary testing on branch
July 24 between it (hack to load from s3 bucket) and standard
torch-xla=2.7.0 and see exact same results, so when that wheel is ready
we can switch to it later.

### What's changed
- Update ModelTester to use backend="tt-experimental" by default, override with backend="tt"
- Add TT_TORCH_FORCE_LEGACY_BACKEND=1 to switch back to legacy tt backend in future
- Some test updates to drop PCC by 0.1 (albert, qwen2), enable consteval (needed in several bigger models)
- Override arg type map with asif change for failing mobilenet ssd and variants
- ModelTester workaround to automatically use legacy tt backend in data_parallel tests to avoid assert (#1105)
- Use legacy tt backend in 4+5+6=15 n300 batch parallel tests hitting L1 OOM errors (#1105)
- Use legacy tt backend in few tests where they failed like inception_v4.tf_in1k (#1104)
- Use legacy tt backend in phi-2 model to avoid host OOM on CI (#1108)
- Remove experimental CI yml files and references, no longer needed with single nightly.
- Remove test_e2e_compile yml usage no longer supported with experimental backend

### Checklist
- [x] A bunch of branch testing on nightly between WH (mostly) and BH (see comments in PR) all pass.

---------

Co-authored-by: James Xu <jameszianxu@tenstorrent.com>