Add --forked to full-model-execute ymls and shuffle high memory phi-2 to high-memory group (#1117)
### Ticket
None

### Problem description
- Tests that use more than 32GB or 48GB memory can be killed with signal
137 which takes down entire pytest group resulting in nothing reported
to superset database.
- We didn't have common `high-memory` tag that could return n150 or n300
high machines for use by tests

### What's changed
- Added deps (pytest-xdist, pytest-forked) to use pytest --forked arg
which gives per-test processes so when tests killed due to OOM, it only
affects a single test and not entire group.
- Move high memory offender phi-2 to it's own group, move previously
quarantined bert test back to common group
- On github settings, tagged all high-memory-n* machines with
`high-memory-wormhole` now and use it here to try and make phi-2 safer
in future runs,

### Checklist
- [x] Tested locally and on branch