Update torch-xla requirements to use tenstorrent/pytorch-xla wheel (#1123)
### Ticket
[794](https://github.com/tenstorrent/tt-xla/issues/794)

### Problem description
We have the *.whl for [tenstorret/torch-xla
](https://github.com/tenstorrent/pytorch-xla) in the tt-pypi server, now
we need to update our requirements to use the wheel. This will allow us
to support muti-chip compilation through torch-xla.

### What's changed
Updated torch-xla install in `requirements.txt` and `setup.py`. Disabled
elementwise binary left/right shift ops due to compilation error. In the
previous iteration of torch-xla the op wasn't being compiled either but
instead of it would fall-back to CPU. We will disable these test for now
and track it in issue #1127.

### Checklist
- [x] New/Existing tests provide coverage for changes