Uplift tt-mlir to b831a7e12e4fc0a75958cd827bbe2bfb2a416353 and tt-xla to e11a4e14fab86fa8695957ac1df663572b80c19c (#1140)
### Ticket
None

### Problem description
Several breaking changes between July 23-30 were introduced in tt-mlir,
introducing build breaks and test failures tt-torch and a build failure
in tt-xla.

### What's changed
We have not uplifted tt-xla and tt-mlir simultaneously before but it is
necessary in this case, because a manual tt-xla uplift is required to
fix the build break introduced by a tt-mlir change. Therefore, to get a
newer tt-mlir, we need to uplift tt-xla at the same time.

Bisect note and follow up changes in issue 1142
https://github.com/tenstorrent/tt-torch/issues/1142

### Checklist
- [x] New/Existing tests provide coverage for changes