Add global sharding registry to allow mark_sharding() to work through torch.compile path (#1189)
### Ticket
Dependency of https://github.com/tenstorrent/tt-xla/issues/1001

### Problem description
It is not currently possible to mark sharding for tensors in models
using the torch.compile flow. This is because we can only mark_sharding
on XLA tensors that have been moved onto device, but we move them onto
device in the XLAbackend where the user cannot access them.

### What's changed
- Create deferred shard marking path using a global sharding registry:
- Users mark sharding by ts.mark_sharding() on raw torch.Tensors to save
them in a global dict. The API for this mirrors the existing torch-xla
API
- In the executor, extract the user shard marking by pulling it out of
the dict
- Use weak references in the global sharding registry to avoid
preventing the original torch tensors from being GC'd
- Create initial pytest for sharding
- Create test fixture to setup and teardown torch-xla SPMD env variables

### Checklist
- [x] New/Existing tests provide coverage for changes